backend/requirements.txt
fastapi>=0.128.0,<0.129.0
uvicorn>=0.24.0,<0.33.0
python-dotenv>=1.0.0,<2.0.0
boto3>=1.34.0,<2.0.0
pydantic>=2.5.0,<3.0.0
supabase>=2.0.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
google-auth>=2.23.4,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.1,<1.0.0
google-generativeai>=0.3.2,<1.0.0
google-cloud-vision>=3.4.0,<4.0.0
google-cloud-storage>=2.10.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
Pillow>=10.1.0,<12.0.0
httpx>=0.25.2,<1.0.0
bcrypt>=4.1.2,<5.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
email-validator>=2.1.0,<3.0.0
cryptography>=41.0.7,<45.0.0
requests>=2.31.0,<3.0.0
slowapi>=0.1.9,<1.0.0
sentry-sdk[fastapi]>=1.39.0,<3.0.0
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<6.0.0
redis>=5.0.0,<6.0.0
python-magic-bin>=0.4.14,<1.0.0
bleach>=6.0.0,<7.0.0
orjson>=3.9.10
# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-exporter-otlp>=1.20.0
